+++ namer/consul/src/test/scala/io/buoyant/namer/consul/ConsulTest.scala
@@ -22,7 +22,6 @@ class ConsulTest extends FunSuite {
   test("parse minimal config") {
     val yaml = s"""
                   |kind: io.l5d.consul
-                  |experimental: true
       """.stripMargin
 
     val mapper = Parser.objectMapper(yaml, Iterable(Seq(ConsulInitializer)))
@@ -35,7 +34,6 @@ class ConsulTest extends FunSuite {
   test("parse all options config") {
     val yaml = s"""
                     |kind: io.l5d.consul
-                    |experimental: true
                     |host: consul.site.biz
                     |port: 8600
                     |token: some-token
@@ -53,19 +51,5 @@ class ConsulTest extends FunSuite {
     assert(consul.includeTag == Some(true))
     assert(consul.consistencyMode == Some(ConsistencyMode.Stale))
     assert(!consul.disabled)
-  }
-
-  test("parse config without experimental param") {
-    val yaml = s"""
-                  |kind: io.l5d.consul
-                  |host: consul.site.biz
-                  |port: 8600
-      """.stripMargin
-
-    val mapper = Parser.objectMapper(yaml, Iterable(Seq(ConsulInitializer)))
-    val consul = mapper.readValue[NamerConfig](yaml).asInstanceOf[ConsulConfig]
-    assert(consul.host == Some("consul.site.biz"))
-    assert(consul.port == Some(Port(8600)))
-    assert(consul.disabled)
   }
 }